+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased (v0.5.2)
 * Fix ffprobe duration conversion error scenarios panicking.
+* Tweak encoded size prediction logic to consider both input file size & encoded sample duration.
 
 # v0.5.1
 * Change encoded size prediction logic to estimate video stream size (or image size) only.